+++ app/Project.php
@@ -31,8 +31,8 @@ class Project extends Model
 
     protected $casts = [
         'company_id' => 'integer',
-        'base_price' => 'integer',
-        'penalty' => 'integer',
+        'base_price' => 'double',
+        'penalty' => 'double',
     ];
 
     /**
