+++ database/migrations/2017_04_04_143804_create_projects_table.php
@@ -16,8 +16,8 @@ class CreateProjectsTable extends Migration
         Schema::create('projects', function (Blueprint $table) {
             $table->increments('id');
             $table->string('name');
-            $table->integer('base_price')->nullable();
-            $table->integer('penalty')->nullable();
+            $table->decimal('base_price', 15, 4)->nullable();
+            $table->decimal('penalty', 15, 4)->nullable();
             $table->timestamps();
             $table->softDeletes();
         });
